Added a new effect and updated the character respectfully.  Need to fix this though.
git-svn-id: svn://10.0.50.55/puffbomb/trunk@2522 d14347b8-48fd-0310-9e84-97907d434d85

+++ Code/Engine2D/DynamicCollection/DynamicCollection.h
@@ -73,7 +73,7 @@ public:
 	// - -------------------------------------------------------------------------------------- - //
 	// Query if this object is, at all, enabled //
 	inline bool IsActive() const {
-		for ( int idx = 0; idx < Component.size(); idx++ ) {
+		for ( size_t idx = 0; idx < Component.size(); idx++ ) {
 			if ( Component[ idx ].IsActive() ) {
 				return true;
 			}
